Some code cleanups for -Wall.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat4@3926 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libGraphics/HandleAxis.cpp
@@ -323,10 +323,10 @@ static QString TrimPrint(double val, bool scientificNotation) {
     while ((label[ePtr] != 'e') && (label[ePtr] != 'E'))
       ePtr--;
     ePtr+=2;
-    while ((label[ePtr] == '0') && ePtr < label.size()) {
+    while ((label[ePtr] == '0') && ePtr < (unsigned int)(label.size())) {
       label.remove(ePtr,1);
     }
-    if (ePtr == label.size())
+    if (ePtr == (unsigned int)(label.size()))
       label.append("0");
     return label;
   }
